+++ src/mock/messaging_api/get_v2_bot_message_aggregation_list.rs
@@ -215,13 +215,12 @@ mod tests {
 
         let captured_req = Arc::new(Mutex::new(Vec::<serde_json::Value>::new()));
         let creq = captured_req.clone();
-        let options = LineOptions {
-            prefix_url: Some(server.url()),
-            ..Default::default()
-        }
-        .with_on_request(move |log| {
-            creq.lock().unwrap().push(log.body().clone());
-        });
+        let options = LineOptions::builder()
+            .with_prefix_url(server.url())
+            .with_on_request(move |log| {
+                creq.lock().unwrap().push(log.body().clone());
+            })
+            .build();
 
         let query_params = get_v2_bot_message_aggregation_list::QueryParams {
             limit: Some(50),
@@ -292,16 +291,15 @@ mod tests {
         let res_count = Arc::new(Mutex::new(0usize));
         let rq = req_count.clone();
         let rs = res_count.clone();
-        let options = LineOptions {
-            prefix_url: Some(server.url()),
-            ..Default::default()
-        }
-        .with_on_request(move |_log| {
-            *rq.lock().unwrap() += 1;
-        })
-        .with_on_response(move |_req, _res| {
-            *rs.lock().unwrap() += 1;
-        });
+        let options = LineOptions::builder()
+            .with_prefix_url(server.url())
+            .with_on_request(move |_log| {
+                *rq.lock().unwrap() += 1;
+            })
+            .with_on_response(move |_req, _res| {
+                *rs.lock().unwrap() += 1;
+            })
+            .build();
 
         let res = get_v2_bot_message_aggregation_list::execute_stream(
             "test_channel_access_token",